lge: display: fixup code, and temporarily remove LGE_BLT_LOCKUP_WR feature
Fixup wrong indents and spaces in the code, and temporarily remove the
LGE_BLT_LOCKUP_WR feature as it is getting in the way. This should be re-added
(of course, with adaptions to the newer code) after patches are applied.
Also, init dsi clocks after MIPI_OUTP for the LGE Iproj (Nitro HD/Optimus LTE)